split out types into seperate file
